commit  promform

+++ src/main/java/commonTasks/dto/FamilleArticleStatDTO.java
@@ -18,9 +18,9 @@ import org.json.JSONPropertyName;
 public class FamilleArticleStatDTO implements Serializable {
     private static final long serialVersionUID = 1L;
     private String code, libelle, familleId, description,id;
-    private Integer montantTTC = 0, montantHT = 0, montantAchat = 0, montantMarge = 0, montantCumulTTC = 0, montantCumulHT = 0, montantCumulAchat = 0;
-    private Integer valeurPeriode = 0, pourcentageMage = 0, pourcentageTH = 0, pourcentageCumulMage = 0, pourcentageCumulTH = 0;
-    private Integer montantTva = 0, montantCumulTva = 0, montantRemise = 0, montantCumulMarge;
+    private long montantTTC = 0, montantHT = 0, montantAchat = 0, montantMarge = 0, montantCumulTTC = 0, montantCumulHT = 0, montantCumulAchat = 0;
+    private long valeurPeriode = 0, pourcentageMage = 0, pourcentageTH = 0, pourcentageCumulMage = 0, pourcentageCumulTH = 0;
+    private long montantTva = 0, montantCumulTva = 0, montantRemise = 0, montantCumulMarge;
 
     public FamilleArticleStatDTO() {
     }
@@ -37,11 +37,11 @@ public class FamilleArticleStatDTO implements Serializable {
         return familleId;
     }
 
-    public Integer getMontantCumulMarge() {
+    public long getMontantCumulMarge() {
         return montantCumulMarge;
     }
 
-    public void setMontantCumulMarge(Integer montantCumulMarge) {
+    public void setMontantCumulMarge(long montantCumulMarge) {
         this.montantCumulMarge = montantCumulMarge;
     }
 
@@ -53,19 +53,19 @@ public class FamilleArticleStatDTO implements Serializable {
         this.description = description;
     }
 
-    public Integer getPourcentageCumulMage() {
+    public long getPourcentageCumulMage() {
         return pourcentageCumulMage;
     }
 
-    public void setPourcentageCumulMage(Integer pourcentageCumulMage) {
+    public void setPourcentageCumulMage(long pourcentageCumulMage) {
         this.pourcentageCumulMage = pourcentageCumulMage;
     }
 
-    public Integer getPourcentageCumulTH() {
+    public long getPourcentageCumulTH() {
         return pourcentageCumulTH;
     }
 
-    public void setPourcentageCumulTH(Integer pourcentageCumulTH) {
+    public void setPourcentageCumulTH(long pourcentageCumulTH) {
         this.pourcentageCumulTH = pourcentageCumulTH;
     }
 
@@ -73,49 +73,49 @@ public class FamilleArticleStatDTO implements Serializable {
         this.familleId = familleId;
     }
 
-    public Integer getMontantTva() {
+    public long getMontantTva() {
         return montantTva;
     }
 
-    public void setMontantTva(Integer montantTva) {
+    public void setMontantTva(long montantTva) {
         this.montantTva = montantTva;
     }
 
-    public Integer getMontantCumulTva() {
+    public long getMontantCumulTva() {
         return montantCumulTva;
     }
 
-    public void setMontantCumulTva(Integer montantCumulTva) {
+    public void setMontantCumulTva(long montantCumulTva) {
         this.montantCumulTva = montantCumulTva;
     }
 
-    public Integer getMontantRemise() {
+    public long getMontantRemise() {
         return montantRemise;
     }
 
-    public void setMontantRemise(Integer montantRemise) {
+    public void setMontantRemise(long montantRemise) {
         this.montantRemise = montantRemise;
     }
 
     public FamilleArticleStatDTO(String code, String libelle, long montantTTC, long montantAchat, long montantTva, long montantRemise, String familleId) {
         this.code = code;
         this.libelle = libelle;
-        this.montantTTC = (int) (montantTTC - montantRemise);
-        this.montantHT = (int) (montantTTC - montantRemise - montantTva);
-        this.montantAchat = (int) montantAchat;
-        this.montantMarge = (int) (this.montantHT - montantAchat);
+        this.montantTTC = (montantTTC - montantRemise);
+        this.montantHT =  (montantTTC - montantRemise - montantTva);
+        this.montantAchat =  montantAchat;
+        this.montantMarge =(this.montantHT - montantAchat);
         Double p = new BigDecimal(Double.valueOf(this.montantMarge) / this.montantHT).setScale(2, RoundingMode.HALF_UP).doubleValue() * 100;
         this.pourcentageMage = p.intValue();
-        this.montantTva = (int) montantTva;
-        this.montantRemise = (int) montantRemise;
+        this.montantTva =  montantTva;
+        this.montantRemise =  montantRemise;
         this.familleId = familleId;
     }
 
     public FamilleArticleStatDTO(long montantTTC, long montantAchat, long montantTva, long montantRemise) {
-        this.montantCumulTTC = (int) (montantTTC - montantRemise);
-        this.montantCumulHT = (int) (montantTTC - montantRemise - montantTva);
-        this.montantCumulAchat = (int) montantAchat;
-        this.montantCumulTva = (int) montantTva;
+        this.montantCumulTTC =  (montantTTC - montantRemise);
+        this.montantCumulHT = (montantTTC - montantRemise - montantTva);
+        this.montantCumulAchat =  montantAchat;
+        this.montantCumulTva =  montantTva;
 
     }
 
@@ -136,88 +136,88 @@ public class FamilleArticleStatDTO implements Serializable {
     }
 
     @JSONPropertyName("MONTANT NET TTC")
-    public Integer getMontantTTC() {
+    public long getMontantTTC() {
         return montantTTC;
     }
 
-    public void setMontantTTC(Integer montantTTC) {
+    public void setMontantTTC(long montantTTC) {
         this.montantTTC = montantTTC;
     }
 
     @JSONPropertyName("MONTANT NET HT")
-    public Integer getMontantHT() {
+    public long getMontantHT() {
         return montantHT;
     }
 
-    public void setMontantHT(Integer montantHT) {
+    public void setMontantHT(long montantHT) {
         this.montantHT = montantHT;
     }
 
     @JSONPropertyName("VALEUR ACHAT")
-    public Integer getMontantAchat() {
+    public long getMontantAchat() {
         return montantAchat;
     }
 
-    public void setMontantAchat(Integer montantAchat) {
+    public void setMontantAchat(long montantAchat) {
         this.montantAchat = montantAchat;
     }
 
     @JSONPropertyName("MARGE NET")
-    public Integer getMontantMarge() {
+    public long getMontantMarge() {
         return montantMarge;
     }
 
-    public void setMontantMarge(Integer montantMarge) {
+    public void setMontantMarge(long montantMarge) {
         this.montantMarge = montantMarge;
     }
 
-    public Integer getMontantCumulTTC() {
+    public long getMontantCumulTTC() {
         return montantCumulTTC;
     }
 
-    public void setMontantCumulTTC(Integer montantCumulTTC) {
+    public void setMontantCumulTTC(long montantCumulTTC) {
         this.montantCumulTTC = montantCumulTTC;
     }
 
-    public Integer getMontantCumulHT() {
+    public long getMontantCumulHT() {
         return montantCumulHT;
     }
 
-    public void setMontantCumulHT(Integer montantCumulHT) {
+    public void setMontantCumulHT(long montantCumulHT) {
         this.montantCumulHT = montantCumulHT;
     }
 
-    public Integer getMontantCumulAchat() {
+    public long getMontantCumulAchat() {
         return montantCumulAchat;
     }
 
-    public void setMontantCumulAchat(Integer montantCumulAchat) {
+    public void setMontantCumulAchat(long montantCumulAchat) {
         this.montantCumulAchat = montantCumulAchat;
     }
 
-    public Integer getValeurPeriode() {
+    public long getValeurPeriode() {
         return valeurPeriode;
     }
 
-    public void setValeurPeriode(Integer valeurPeriode) {
+    public void setValeurPeriode(long valeurPeriode) {
         this.valeurPeriode = valeurPeriode;
     }
 
     @JSONPropertyName("MARGE POURCENTAGE")
-    public Integer getPourcentageMage() {
+    public long getPourcentageMage() {
         return pourcentageMage;
     }
 
-    public void setPourcentageMage(Integer pourcentageMage) {
+    public void setPourcentageMage(long pourcentageMage) {
         this.pourcentageMage = pourcentageMage;
     }
 
     @JSONPropertyName("POURCENTAGE TOTAL")
-    public Integer getPourcentageTH() {
+    public long getPourcentageTH() {
         return pourcentageTH;
     }
 
-    public void setPourcentageTH(Integer pourcentageTH) {
+    public void setPourcentageTH(long pourcentageTH) {
         this.pourcentageTH = pourcentageTH;
     }
 
@@ -246,18 +246,18 @@ public class FamilleArticleStatDTO implements Serializable {
         return true;
     }
 
-    public FamilleArticleStatDTO(String code, String libelle, long montantTTC, long montantAchat, long montantTva, long montantRemise, Integer prixAchat, Integer prixVente, long quantite) {
+     public FamilleArticleStatDTO(String code, String libelle, long montantTTC, long montantAchat, long montantTva, long montantRemise, Integer prixAchat, Integer prixVente, long quantite){
         this.code = code;
         this.libelle = libelle;
-        this.montantCumulTTC = (int) (montantTTC - montantRemise);
-        this.montantCumulHT = (int) (montantTTC - montantRemise - montantTva);
-        this.montantCumulAchat = (int) montantAchat;
-        this.montantCumulMarge = (int) (this.montantCumulHT - montantAchat);
+        this.montantCumulTTC = (montantTTC - montantRemise);
+        this.montantCumulHT = (montantTTC - montantRemise - montantTva);
+        this.montantCumulAchat =  montantAchat;
+        this.montantCumulMarge =  (this.montantCumulHT - montantAchat);
         Double p = new BigDecimal(Double.valueOf(this.montantCumulMarge) / this.montantCumulHT).setScale(2, RoundingMode.HALF_UP).doubleValue() * 100;
         this.pourcentageCumulMage = p.intValue();
         this.montantTva = prixAchat;
         this.montantRemise = prixVente;
-        this.montantCumulTva = (int) quantite;
+        this.montantCumulTva =  quantite;
 
     }
 /**
@@ -276,32 +276,32 @@ public class FamilleArticleStatDTO implements Serializable {
  */
   
     public FamilleArticleStatDTO(String id,String code, String libelle, String familleId, String description, long quantiteVNO, long quantiteVO, long quantiteVendue,
-            long nombreSortie, long montantVente,Integer seuil) {
+            long nombreSortie, long montantVente,int seuil) {
         this.code = code;
         this.libelle = libelle;
         this.familleId = familleId;
         this.description = description;
-        this.montantCumulTTC = (int) quantiteVNO;
-        this.montantCumulHT = (int) quantiteVO;
-        this.montantCumulAchat = (int) quantiteVendue;
-        this.montantCumulMarge = (int) nombreSortie;
+        this.montantCumulTTC =  quantiteVNO;
+        this.montantCumulHT = quantiteVO;
+        this.montantCumulAchat =  quantiteVendue;
+        this.montantCumulMarge = nombreSortie;
         Double p = new BigDecimal(Double.valueOf(quantiteVendue) / nombreSortie).setScale(1, RoundingMode.HALF_UP).doubleValue();
         this.pourcentageCumulMage = p.intValue();
-        this.montantTva = (int) montantVente;
-        this.valeurPeriode=seuil;
+        this.montantTva =  montantVente;
+        this.valeurPeriode= Long.valueOf(seuil) ;
         this.id=id;
 
     }
   public FamilleArticleStatDTO(String code, String libelle,String familleId, String description, long montantTTC, long montantAchat, long montantTva, long montantRemise, long quantite) {
         this.code = code;
         this.libelle = libelle;
-        this.montantCumulTTC = (int) (montantTTC - montantRemise);
-        this.montantCumulHT = (int) (montantTTC - montantRemise - montantTva);
-        this.montantCumulAchat = (int) montantAchat;
-        this.montantCumulMarge = (int) (this.montantCumulHT - montantAchat);
+        this.montantCumulTTC =  (montantTTC - montantRemise);
+        this.montantCumulHT =  (montantTTC - montantRemise - montantTva);
+        this.montantCumulAchat =  montantAchat;
+        this.montantCumulMarge =  (this.montantCumulHT - montantAchat);
         Double p = new BigDecimal(Double.valueOf(this.montantCumulMarge) / this.montantCumulHT).setScale(2, RoundingMode.HALF_UP).doubleValue() * 100;
         this.pourcentageCumulMage = p.intValue();
-        this.montantCumulTva = (int) quantite;
+        this.montantCumulTva =  quantite;
          this.familleId = familleId;
         this.description = description;
